Refactor backend tests
