+++ backend/tests_back/test_chatbot.py
@@ -214,279 +214,4 @@ async def test_stream_chat_error_handling(chatbot_service):
     assert "Stream error" in chunks[0]
 
 
-@pytest.fixture
-def test_db(mocker):
-    """Create test database"""
-    # Mock the DA_DB_URL environment variable
-    mocker.patch.dict(os.environ, {"DA_DB_URL": "sqlite:///:memory:"})
 
-    from app.db.models import Base
-    from sqlalchemy import create_engine
-    from sqlalchemy.orm import sessionmaker
-    from sqlalchemy.pool import StaticPool
-
-    test_engine = create_engine(
-        "sqlite:///:memory:",
-        connect_args={"check_same_thread": False},
-        poolclass=StaticPool,
-    )
-    Base.metadata.create_all(bind=test_engine)
-    TestingSessionLocal = sessionmaker(
-        autocommit=False, autoflush=False, bind=test_engine
-    )
-    return TestingSessionLocal
-
-
-@pytest.fixture
-def test_client(test_db, mocker):
-    """Create test client with mocked dependencies"""
-    # Mock environment variables before importing app
-    mocker.patch.dict(
-        os.environ,
-        {
-            "DA_OPENAI_API_KEY": "test-key-123",
-            "OPENAI_API_KEY": "test-key-123",
-        },
-    )
-
-    from fastapi.testclient import TestClient
-    from app.main import app, get_db, get_user_id
-
-    def override_get_db():
-        db = test_db()
-        try:
-            yield db
-        finally:
-            db.close()
-
-    def override_get_user_id():
-        return "test_user_123"
-
-    app.dependency_overrides[get_db] = override_get_db
-    app.dependency_overrides[get_user_id] = override_get_user_id
-
-    client = TestClient(app)
-    yield client
-
-    app.dependency_overrides.clear()
-
-
-def test_save_conversation(test_client):
-    """Test saving a conversation with messages"""
-    conversation_data = {
-        "conversation_starter": "Hello, how are you?",
-        "thread_id": "test-thread-001",
-        "model": "gpt-4o",
-        "system_prompt_a": "You are a helpful assistant.",
-        "system_prompt_b": "You are a knowledgeable expert.",
-        "turns": 3,
-        "messages": [
-            {"chatbot": "user", "message": "Hello, how are you?"},
-            {"chatbot": "a", "message": "I'm doing well, thank you!"},
-            {"chatbot": "b", "message": "That's great to hear!"},
-        ],
-    }
-
-    response = test_client.post("/conversations", json=conversation_data)
-
-    assert response.status_code == 200
-    data = response.json()
-
-    assert data["user"] == "test_user_123"
-    assert data["conversation_starter"] == "Hello, how are you?"
-    assert data["thread_id"] == "test-thread-001"
-    assert data["model"] == "gpt-4o"
-    assert data["system_prompt_a"] == "You are a helpful assistant."
-    assert data["system_prompt_b"] == "You are a knowledgeable expert."
-    assert data["turns"] == 3
-    assert "id" in data
-    assert "created_at" in data
-
-    assert len(data["messages"]) == 3
-    assert data["messages"][0]["chatbot"] == "user"
-    assert data["messages"][0]["message"] == "Hello, how are you?"
-    assert data["messages"][0]["order"] == 0
-    assert data["messages"][1]["chatbot"] == "a"
-    assert data["messages"][1]["order"] == 1
-    assert data["messages"][2]["chatbot"] == "b"
-    assert data["messages"][2]["order"] == 2
-
-
-def test_retrieve_saved_conversation(test_client):
-    """Test retrieving a saved conversation"""
-    conversation_data = {
-        "conversation_starter": "Test message",
-        "thread_id": "test-thread-002",
-        "turns": 2,
-        "messages": [
-            {"chatbot": "user", "message": "Test message"},
-            {"chatbot": "a", "message": "Test response"},
-        ],
-    }
-
-    post_response = test_client.post("/conversations", json=conversation_data)
-    conversation_id = post_response.json()["id"]
-
-    get_response = test_client.get(f"/conversations/{conversation_id}")
-
-    assert get_response.status_code == 200
-    data = get_response.json()
-    assert data["id"] == conversation_id
-    assert len(data["messages"]) == 2
-
-
-def test_save_multiple_conversations(test_client):
-    """Test saving multiple conversations"""
-    conv1_data = {
-        "conversation_starter": "First question",
-        "thread_id": "thread-1",
-        "model": "gpt-4o",
-        "turns": 2,
-        "messages": [
-            {"chatbot": "user", "message": "First question"},
-            {"chatbot": "a", "message": "First answer"},
-        ],
-    }
-
-    conv2_data = {
-        "conversation_starter": "Second question",
-        "thread_id": "thread-2",
-        "model": "gpt-4o-mini",
-        "turns": 3,
-        "messages": [
-            {"chatbot": "user", "message": "Second question"},
-            {"chatbot": "a", "message": "Second answer"},
-            {"chatbot": "b", "message": "More details"},
-        ],
-    }
-
-    response1 = test_client.post("/conversations", json=conv1_data)
-    response2 = test_client.post("/conversations", json=conv2_data)
-
-    assert response1.status_code == 200
-    assert response2.status_code == 200
-
-    list_response = test_client.get("/conversations")
-
-    assert list_response.status_code == 200
-    conversations = list_response.json()
-    assert len(conversations) == 2
-
-
-@pytest.mark.parametrize(
-    "turns,expected_status",
-    [
-        (0, 422),
-        (25, 422),
-        (10, 200),
-        (1, 200),
-        (20, 200),
-    ],
-)
-def test_conversation_validation_turns(test_client, turns, expected_status):
-    """Test conversation validation with various turn values"""
-    data = {
-        "conversation_starter": "Test",
-        "thread_id": f"thread-test-{turns}",
-        "turns": turns,
-        "messages": [{"chatbot": "user", "message": "Test"}],
-    }
-
-    response = test_client.post("/conversations", json=data)
-    assert response.status_code == expected_status
-
-
-def test_delete_conversation(test_client):
-    """Test deleting a conversation"""
-    conv_data = {
-        "conversation_starter": "To be deleted",
-        "thread_id": "thread-delete",
-        "turns": 2,
-        "messages": [
-            {"chatbot": "user", "message": "To be deleted"},
-            {"chatbot": "a", "message": "Response"},
-        ],
-    }
-
-    create_response = test_client.post("/conversations", json=conv_data)
-    assert create_response.status_code == 200
-    conversation_id = create_response.json()["id"]
-
-    delete_response = test_client.delete(f"/conversations/{conversation_id}")
-
-    assert delete_response.status_code == 200
-    assert delete_response.json()["message"] == "Conversation deleted successfully"
-
-    get_response = test_client.get(f"/conversations/{conversation_id}")
-    assert get_response.status_code == 404
-
-
-def test_delete_nonexistent_conversation(test_client):
-    """Test deleting a non-existent conversation"""
-    response = test_client.delete("/conversations/99999")
-
-    assert response.status_code == 404
-    data = response.json()
-    assert "detail" in data
-
-
-def test_save_and_get_prompts(test_client):
-    """Save a prompt and retrieve it via /get_prompts."""
-    payload = {"agent_name": "agent1", "prompt": "This is a test prompt"}
-    r = test_client.post("/save_prompt", json=payload)
-    assert r.status_code == 200
-    data = r.json()
-    assert data["agent_name"] == "agent1"
-
-    get_r = test_client.get("/get_prompts")
-    assert get_r.status_code == 200
-    prompts = get_r.json()
-    assert any(p["agent_name"] == "agent1" for p in prompts)
-
-
-def test_save_duplicate_agent_name_returns_409(test_client):
-    """Saving two prompts with same agent_name should return 409 on second save."""
-    payload = {"agent_name": "dup", "prompt": "p"}
-    r1 = test_client.post("/save_prompt", json=payload)
-    assert r1.status_code == 200
-    r2 = test_client.post("/save_prompt", json=payload)
-    assert r2.status_code == 409
-
-
-def test_update_prompt_and_conflicts(test_client):
-    """Update a prompt and ensure name conflicts are handled."""
-    p1 = {"agent_name": "one", "prompt": "p1"}
-    p2 = {"agent_name": "two", "prompt": "p2"}
-    r1 = test_client.post("/save_prompt", json=p1)
-    r2 = test_client.post("/save_prompt", json=p2)
-    assert r1.status_code == 200
-    assert r2.status_code == 200
-
-    id2 = r2.json()["id"]
-
-    # Attempt to rename second prompt to the first's name -> should conflict
-    conflict_payload = {"agent_name": "one", "prompt": "changed"}
-    r_conflict = test_client.put(f"/update_prompt/{id2}", json=conflict_payload)
-    assert r_conflict.status_code == 409
-
-    # Update second prompt to a new unique name -> should succeed
-    update_payload = {"agent_name": "two-new", "prompt": "updated"}
-    r_update = test_client.put(f"/update_prompt/{id2}", json=update_payload)
-    assert r_update.status_code == 200
-    assert r_update.json()["agent_name"] == "two-new"
-
-
-def test_delete_prompt(test_client):
-    """Delete a prompt and ensure it's removed."""
-    payload = {"agent_name": "delme", "prompt": "please delete"}
-    r = test_client.post("/save_prompt", json=payload)
-    assert r.status_code == 200
-    pid = r.json()["id"]
-
-    d = test_client.delete(f"/delete_prompt/{pid}")
-    assert d.status_code == 200
-
-    # Ensure prompt is gone
-    get_r = test_client.get("/get_prompts")
-    assert get_r.status_code == 200
-    assert all(p["id"] != pid for p in get_r.json())
